Context refactoring
