+++ .env
@@ -1,0 +1,1 @@
+ VITE_API_KEY=b168952b
