+++ src/shared/api/types.ts
@@ -3,3 +3,36 @@ export interface SearchParams {
   year?: string;
   plot?: 'short' | 'full';
 }
+
+export interface MovieRatings {
+  Source: string;
+  Value: string;
+}
+
+export interface MoviesResponse {
+  Title: string;
+  Year: string;
+  Rated: string;
+  Released: string;
+  Runtime: string;
+  Genre: string;
+  Director: string;
+  Writer: string;
+  Actors: string;
+  Plot: string;
+  Language: string;
+  Country: string;
+  Awards: string;
+  Poster: string;
+  Ratings: MovieRatings[];
+  Metascore: string;
+  imdbRating: string;
+  imdbVotes: string;
+  imdbID: string;
+  Type: string;
+  DVD: string;
+  BoxOffice: string;
+  Production: string;
+  Website: string;
+  Response: string;
+}
